+++ gcal/gcal_sync/doctype/sync_configuration/sync_configuration.json
@@ -2,12 +2,21 @@
  "allow_copy": 0, 
  "allow_import": 0, 
  "allow_rename": 0, 
+ "autoname": "field:gmail_id", 
  "creation": "2015-06-12 15:24:46.001844", 
  "custom": 0, 
  "docstatus": 0, 
  "doctype": "DocType", 
  "document_type": "Transaction", 
  "fields": [
+  {
+   "fieldname": "gmail_id", 
+   "fieldtype": "Data", 
+   "label": "Gmail ID", 
+   "permlevel": 0, 
+   "precision": "", 
+   "reqd": 1
+  }, 
   {
    "allow_on_submit": 0, 
    "fieldname": "is_sync", 
@@ -136,7 +145,7 @@
  "is_submittable": 0, 
  "issingle": 0, 
  "istable": 0, 
- "modified": "2015-06-16 16:30:55.240438", 
+ "modified": "2015-06-16 17:00:32.968809", 
  "modified_by": "Administrator", 
  "module": "GCal Sync", 
  "name": "Sync Configuration",
