+++ gcal/gcal_sync/doctype/sync_configuration/sync_configuration.py
@@ -100,7 +100,8 @@ def sync_calender():
 		url = get_oauth2_authorize_url('gcal')
 		return url
 	else:
-		sync_events(credentials)
+		from gcal.tasks import sync_google_calendar
+		sync_google_calendar(credentials)
 		return None
 
 @frappe.whitelist()
@@ -127,72 +128,73 @@ def get_credentials(code):
 	frappe.local.response["type"] = "redirect"
 	frappe.local.response["location"] = "/desk#Calendar/Event"
 
-def sync_events(credentials):
-	events = get_gcal_events(credentials)
-	if not events:
-		frappe.msgprint("No Events to Sync")
-	else:
-		for event in events:
-			# check if event alreay synced if exist update else create new event
-			e_name = is_event_already_exist(event)
-			update_event(e_name, event) if e_name else save_event(event)
+# def sync_events(credentials):
+# 	events = get_gcal_events(credentials)
+# 	if not events:
+# 		frappe.msgprint("No Events to Sync")
+# 	else:
+# 		for event in events:
+# 			# check if event alreay synced if exist update else create new event
+# 			e_name = is_event_already_exist(event)
+# 			update_event(e_name, event) if e_name else save_event(event)
 
-def get_gcal_events(credentials):
-	now = datetime.utcnow().isoformat() + 'Z' # 'Z' indicates UTC time
-	service = build('calendar', 'v3', http=credentials.authorize(Http()))
-	eventsResult = service.events().list(
-		calendarId='primary', timeMin=now, maxResults=10, singleEvents=True,
-		orderBy='startTime').execute()
-	events = eventsResult.get('items', [])
+# def get_gcal_events(credentials):
+# 	now = datetime.utcnow().isoformat() + 'Z' # 'Z' indicates UTC time
+# 	service = build('calendar', 'v3', http=credentials.authorize(Http()))
+# 	eventsResult = service.events().list(
+# 		calendarId='primary', timeMin=now, maxResults=10, singleEvents=True,
+# 		orderBy='startTime').execute()
+# 	events = eventsResult.get('items', [])
 
-	return events
+# 	return events
 
-def save_event(event):
-	e = frappe.new_doc("Event")
-	e = set_values(e, event)
-	e.save(ignore_permissions=True)
-	frappe.db.commit()
+# def save_event(event):
+# 	e = frappe.new_doc("Event")
+# 	e = set_values(e, event)
+# 	e.save(ignore_permissions=True)
+# 	frappe.db.commit()
 
-def update_event(name, event):
-	e = frappe.get_doc("Event", name)
-	e = set_values(e, event)
-	e.save(ignore_permissions=True)
-	frappe.db.commit()
+# def update_event(name, event):
+# 	e = frappe.get_doc("Event", name)
+# 	e = set_values(e, event)
+# 	e.save(ignore_permissions=True)
+# 	frappe.db.commit()
 
-def set_values(doc, event):
-	doc.subject = event.get('summary')
+# def set_values(doc, event):
+# 	doc.subject = event.get('summary')
 
-	start_date = event['start'].get('dateTime', event['start'].get('date'))
-	end_date = event['end'].get('dateTime', event['start'].get('date'))
+# 	start_date = event['start'].get('dateTime', event['start'].get('date'))
+# 	end_date = event['end'].get('dateTime', event['start'].get('date'))
 
-	doc.starts_on = get_formatted_date(start_date)
-	doc.ends_on = get_formatted_date(end_date)
+# 	doc.starts_on = get_formatted_date(start_date)
+# 	doc.ends_on = get_formatted_date(end_date)
 	
-	doc.all_day = 1 if doc.starts_on == doc.ends_on else 0
+# 	doc.all_day = 1 if doc.starts_on == doc.ends_on else 0
 
-	if not event.get('visibility'):
-		doc.event_type = "Private"
-	else:
-		doc.event_type =  "Private" if event['visibility'] == "private" else "Public"
+# 	if not event.get('visibility'):
+# 		doc.event_type = "Private"
+# 	else:
+# 		doc.event_type =  "Private" if event['visibility'] == "private" else "Public"
 
-	doc.description = event.get("description")
-	doc.is_gcal_event = 1
-	doc.event_owner = event.get("organizer").get("email")
-	doc.gcal_id = event.get("id")
+# 	doc.description = event.get("description")
+# 	doc.is_gcal_event = 1
+# 	doc.event_owner = event.get("organizer").get("email")
+# 	doc.gcal_id = event.get("id")
 
-	return doc
+# 	return doc
 
-def get_formatted_date(str_date):
-	# Also format the date according to frappe date format
-	date_list = str_date.split("T")
-	date = None
+# def get_formatted_date(str_date):
+# 	# Also format the date according to frappe date format
+# 	date_list = str_date.split("T")
+# 	date = None
 	
-	if len(date_list) == 1:
-		str_date = date_list[0] + "T00:00:00"
+# 	if len(date_list) == 1:
+# 		str_date = date_list[0] + "T00:00:00"
 	
-	date = datetime.strptime(str_date, '%Y-%m-%dT%H:%M:%S').strftime("%d-%m-%Y %H:%M:%S")
-	return datetime.strptime(date, "%d-%m-%Y %H:%M:%S")
+# 	date = datetime.strptime(str_date, '%Y-%m-%dT%H:%M:%S').strftime("%d-%m-%Y %H:%M:%S")
+# 	# return datetime.strptime(date, "%d-%m-%Y %H:%M:%S")
+# 	return date
 
-def is_event_already_exist(event):
-	name = frappe.db.get_value("Event",{"gcal_id":event.get("id")},"name")
-	return name
+# def is_event_already_exist(event):
+# 	name = frappe.db.get_value("Event",{"gcal_id":event.get("id")},"name")
+# 	return name
